Changed Double to decimal for the total pounds value

+++ Money Credit Counter/Form1.cs
@@ -15,7 +15,7 @@ namespace Money_Credit_Counter
     {
 
         int Total_pence = 0;
-        Double Total_Pound = 0.00;
+        decimal Total_Pound = 0.00m;
         int count_1p = 0;
         int count_2p = 0;
         int count_5p = 0;
@@ -48,7 +48,7 @@ namespace Money_Credit_Counter
             {
 
                 Total_pence += 1; // adds 1p to the total pence value
-                Total_Pound += 0.01; // adds 1p to the total pound value 
+                Total_Pound += 0.01m; // adds 1p to the total pound value 
                 count_1p += 1; // adds 1 to the count of the coin
             }
             lbl_1p.Text = count_1p.ToString(); // displays the count on the label for the coin
@@ -69,7 +69,7 @@ namespace Money_Credit_Counter
             {
 
                 Total_pence += 2; // adds 2p to the total pence value
-                Total_Pound += 0.02; // adds 2p to the total pound value 
+                Total_Pound += 0.02m; // adds 2p to the total pound value 
                 count_2p += 1; // adds 1 to the count of the coin
             }
             lbl_2p.Text = count_2p.ToString(); // displays the count on the label for the coin
@@ -94,7 +94,7 @@ namespace Money_Credit_Counter
             {
 
                 Total_pence += 5; // adds 5p to the total pence value
-                Total_Pound += 0.05; // adds 5p to the total pound value 
+                Total_Pound += 0.05m; // adds 5p to the total pound value 
                 count_5p += 1; // adds 1 to the count of the coin
             }
             lbl_5p.Text = count_5p.ToString(); // displays the count on the label for the coin
@@ -112,7 +112,7 @@ namespace Money_Credit_Counter
             {
 
                 Total_pence += 10; // adds 10p to the total pence value
-                Total_Pound += 0.10; // adds 10p to the total pound value 
+                Total_Pound += 0.10m; // adds 10p to the total pound value 
                 count_10p += 1; // adds 1 to the count of the coin
             }
             lbl_10p.Text = count_10p.ToString(); // displays the count on the label for the coin
@@ -130,7 +130,7 @@ namespace Money_Credit_Counter
             {
 
                 Total_pence += 20; // adds 20p to the total pence value
-                Total_Pound += 0.20; // adds 20p to the total pound value 
+                Total_Pound += 0.20m; // adds 20p to the total pound value 
                 count_20p += 1; // adds 1 to the count of the coin
             }
             lbl_20p.Text = count_20p.ToString(); // displays the count on the label for the coin
@@ -148,7 +148,7 @@ namespace Money_Credit_Counter
             {
 
                 Total_pence += 50; // adds 50p to the total pence value
-                Total_Pound += 0.50; // adds 50p to the total pound value 
+                Total_Pound += 0.50m; // adds 50p to the total pound value 
                 count_50p += 1; // adds 1 to the count of the coin
             }
             lbl_50p.Text = count_50p.ToString(); // displays the count on the label for the coin
@@ -165,7 +165,7 @@ namespace Money_Credit_Counter
             {
 
                 Total_pence += 100; // adds 100p to the total pence value
-                Total_Pound += 1.00; // adds £1 to the total pound value 
+                Total_Pound += 1.00m; // adds £1 to the total pound value 
                 count_1pound += 1; // adds 1 to the count of the coin
             }
             lbl_1pound.Text = count_1pound.ToString(); // displays the count on the label for the coin
@@ -182,7 +182,7 @@ namespace Money_Credit_Counter
             {
 
                 Total_pence += 200; // adds 200p to the total pence value
-                Total_Pound += 2.00; // adds £2 to the total pound value 
+                Total_Pound += 2.00m; // adds £2 to the total pound value 
                 count_2pound += 1; // adds 1 to the count of the coin
             }
             lbl_2pound.Text = count_2pound.ToString(); // displays the count on the label for the coin
